requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/requirements.in
#
amqp==5.2.0
    # via kombu
asgiref==3.8.1
    # via
    #   django
    #   django-browser-reload
    #   django-htmx
asttokens==2.4.1
    # via stack-data
billiard==4.2.0
    # via celery
black==24.4.2
    # via -r requirements/requirements.in
bleach[css]==5.0.1
    # via django-bleach
boto3==1.34.149
    # via -r requirements/requirements.in
botocore==1.34.149
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
celery==5.4.0
    # via -r requirements/requirements.in
certifi==2024.7.4
    # via
    #   requests
    #   sentry-sdk
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   djlint
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via djlint
confusable-homoglyphs==3.3.1
    # via django-registration
coverage[toml]==7.6.0
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
cssbeautifier==1.15.1
    # via djlint
decorator==5.1.1
    # via ipython
distlib==0.3.8
    # via virtualenv
dj-database-url==2.2.0
    # via
    #   -r requirements/requirements.in
    #   environs
dj-email-url==1.0.6
    # via environs
django==5.0.7
    # via
    #   -r requirements/requirements.in
    #   dj-database-url
    #   django-bleach
    #   django-browser-reload
    #   django-debug-toolbar
    #   django-extensions
    #   django-honeypot
    #   django-htmx
    #   django-registration
    #   django-storages
    #   model-bakery
django-bleach==3.1.0
    # via -r requirements/requirements.in
django-browser-reload==1.13.0
    # via -r requirements/requirements.in
django-cache-url==3.4.5
    # via environs
django-debug-toolbar==4.4.6
    # via -r requirements/requirements.in
django-extensions==3.2.3
    # via -r requirements/requirements.in
django-honeypot==1.2.0
    # via -r requirements/requirements.in
django-htmx==1.18.0
    # via -r requirements/requirements.in
django-registration==3.4
    # via -r requirements/requirements.in
django-storages==1.14.4
    # via -r requirements/requirements.in
djlint==1.34.1
    # via -r requirements/requirements.in
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
environs[django]==11.0.0
    # via -r requirements/requirements.in
execnet==2.1.1
    # via pytest-xdist
executing==2.0.1
    # via stack-data
faker==26.0.0
    # via -r requirements/requirements.in
filelock==3.15.4
    # via virtualenv
flake8==7.1.0
    # via -r requirements/requirements.in
gunicorn==22.0.0
    # via -r requirements/requirements.in
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
huey==2.5.1
    # via -r requirements/requirements.in
identify==2.6.0
    # via pre-commit
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
ipython==8.26.0
    # via -r requirements/requirements.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.25
    # via djlint
kombu==5.3.7
    # via celery
markdown==3.6
    # via -r requirements/requirements.in
marshmallow==3.21.3
    # via environs
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
model-bakery==1.18.2
    # via -r requirements/requirements.in
mypy-extensions==1.0.0
    # via black
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via
    #   black
    #   build
    #   gunicorn
    #   marshmallow
    #   pytest
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via
    #   black
    #   djlint
pexpect==4.9.0
    # via ipython
pillow==10.4.0
    # via -r requirements/requirements.in
pillow-avif-plugin==1.4.6
    # via -r requirements/requirements.in
pip-lock==2.11.0
    # via -r requirements/requirements.in
pip-tools==7.4.1
    # via -r requirements/requirements.in
platformdirs==4.2.2
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.7.1
    # via -r requirements/requirements.in
prompt-toolkit==3.0.47
    # via
    #   click-repl
    #   ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pycodestyle==2.12.0
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-xdist
pytest-cov==5.0.0
    # via -r requirements/requirements.in
pytest-django==4.8.0
    # via -r requirements/requirements.in
pytest-xdist==3.6.1
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   faker
python-dotenv==1.0.1
    # via environs
pytz==2024.1
    # via -r requirements/requirements.in
pyyaml==6.0.1
    # via
    #   djlint
    #   pre-commit
regex==2023.12.25
    # via djlint
requests[security]==2.31.0
    # via -r requirements/requirements.in
s3transfer==0.10.2
    # via boto3
sentry-sdk==2.11.0
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   cssbeautifier
    #   jsbeautifier
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
tinycss2==1.1.1
    # via bleach
tqdm==4.66.4
    # via djlint
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via dj-database-url
tzdata==2024.1
    # via celery
urllib3==2.2.2
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.26.3
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.43.0
    # via pip-tools
whitenoise==6.7.0
    # via -r requirements/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
